Improve per log filter storage size

+++ src-tauri/src/controller/common/index_range.rs
@@ -7,7 +7,7 @@
 //! # `index.range.rs`
 //!
 //! **Author**: Alexandru Delegeanu
-//! **Version**: 0.2
+//! **Version**: 0.3
 //! **Description**: Utility struct. Inclusive range [begin, end]
 //!
 
@@ -15,28 +15,28 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct IndexRange {
-    begin: usize,
-    end: usize,
+    begin: u64,
+    end: u64,
 }
 
 impl IndexRange {
-    pub fn new(begin: usize, end: usize) -> Self {
+    pub fn new(begin: u64, end: u64) -> Self {
         Self { begin, end }
     }
 
-    pub fn contains(&self, index: usize) -> bool {
+    pub fn contains(&self, index: u64) -> bool {
         self.begin <= index && index <= self.end
     }
 
     pub fn size(&self) -> usize {
-        self.end - self.begin + 1
+        (self.end - self.begin + 1) as usize
     }
 
-    pub fn begin(&self) -> usize {
+    pub fn begin(&self) -> u64 {
         self.begin
     }
 
-    pub fn end(&self) -> usize {
+    pub fn end(&self) -> u64 {
         self.end
     }
 }
